13_docker/04_frontend-backend-db-app/backend/requirements.txt
Flask==2.3.2
psycopg2-binary==2.9.9
gunicorn==22.0.0
Flask-CORS==4.0.1
